@ElizaWszola 2024-03-07
@zhuohan123 I ran `benchmark_throughput.py` with and without auto prefix caching enabled, and it looks like the code is slightly faster after the fix:

old version with auto prefix caching
```Throughput: 8.61 requests/s, 4166.47 tokens/s```

new version with auto prefix caching
```Throughput: 8.92 requests/s, 4317.72 tokens/s```

old version without auto prefix caching
```Throughput: 8.98 requests/s, 4347.19 tokens/s```

new version without auto prefix caching
```Throughput: 9.10 requests/s, 4401.76 tokens/s```